% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv_indel_helpers.r
\name{add_colour_vector_to_concordance}
\alias{add_colour_vector_to_concordance}
\title{Add colour vector to concordance data frame}
\usage{
add_colour_vector_to_concordance()
}
\description{
This is a dummy function - this function is defined in the upstream analysis documents
}
