% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_voom_for_feature.r
\name{add_entrez_ids_from_hugo_genes}
\alias{add_entrez_ids_from_hugo_genes}
\title{Add Entrez gene IDs to a data frame of voom hits}
\usage{
add_entrez_ids_from_hugo_genes(voom_hits)
}
\arguments{
\item{voom_hits}{A data frame containing voom hits, with HUGO gene ids}
}
\value{
df A data frame containing the Entrez gene IDs in a column called 'entrez'
}
\description{
Add Entrez gene IDs to a data frame of voom hits
}
