% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratification_functions.r
\name{apply_eln2015_cyto}
\alias{apply_eln2015_cyto}
\title{Apply ELN2015 stratification based on cytogenetics}
\usage{
apply_eln2015_cyto(cyto_status, npm1, flt3_itd, cebpa)
}
\arguments{
\item{cyto_status}{Cytogenetic status}

\item{npm1}{NPM1 status}

\item{flt3_itd}{FLT3-ITD status}

\item{cebpa}{CEBPA status}
}
\value{
A stratification status and rationale
}
\description{
Apply ELN2015 stratification based on cytogenetics
}
