% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratification_functions.r
\name{apply_eln2017_cyto}
\alias{apply_eln2017_cyto}
\title{Apply ELN2017 stratification based on cytogenetics}
\usage{
apply_eln2017_cyto(cyto_status, npm1, flt3_itd, flt3_itd_support, cebpa,
  tp53, runx1, asxl1)
}
\arguments{
\item{cyto_status}{Cytogenetic status}

\item{npm1}{NPM1 status}

\item{flt3_itd}{FLT3-ITD status}

\item{flt3_itd_support}{FLT3-ITD support (high or low)}

\item{cebpa}{CEBPA status}

\item{tp53}{TP53 status}

\item{runx1}{RUNX1 status}

\item{asxl1}{ASXL1 status}
}
\value{
A stratification status and rationale
}
\description{
Apply ELN2017 stratification based on cytogenetics
}
