% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratification_functions.r
\name{apply_expression_signature_reclassification}
\alias{apply_expression_signature_reclassification}
\title{Update an existing stratification by applying a gene expression score}
\usage{
apply_expression_signature_reclassification(original_cat, expression_score,
  fav_threshold, adverse_threshold)
}
\arguments{
\item{original_cat}{The original stratification category}

\item{expression_score}{The gene expression signature score}

\item{adverse_threshold}{Threshold for adjusting stratification to 'adverse' category}

\item{fav_theshold}{Threshold for adjusting stratification to 'favourable' category}
}
\value{
A revised stratification category and justification
}
\description{
Update an existing stratification by applying a gene expression score
}
