% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.r
\name{aps_volcano_plot}
\alias{aps_volcano_plot}
\title{Draw Volcano Plot for APS DE Comparison}
\usage{
aps_volcano_plot(
  df,
  q_threshold,
  fc_threshold,
  label_q_threshold,
  label_fc_threshold,
  draw_labels = FALSE
)
}
\arguments{
\item{df}{Data-frame containing DE results.}

\item{q_threshold}{Q-threshold for colouring points.}

\item{fc_threshold}{Fold-change threshold for colouring points.}

\item{label_q_threshold}{Q-value threshold for adding labels.}

\item{label_fc_threshold}{Fold-change threshold for adding labels.}

\item{draw_labels}{Boolean - add labels?}
}
\value{
A ggplot object
}
\description{
Draw Volcano Plot for APS DE Comparison
}
