% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cna_workflow.r
\name{calculate_reference_set_mean_expression}
\alias{calculate_reference_set_mean_expression}
\title{Calculate reference set mean expression}
\usage{
calculate_reference_set_mean_expression(expr.mat, reference_set)
}
\arguments{
\item{reference_set}{}
}
\description{
In this case, we have a matrix of expression values, and a character vector of samples in the reference set
What we want to do is calculate the rowMeans from just the samples in the reference set
}
