% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.r
\name{connect_to_aml_project_orm}
\alias{connect_to_aml_project_orm}
\title{Connect to AML Project ORM Database}
\usage{
connect_to_aml_project_orm(dbname = "aml_project_db",
  host = "karsanlab-db01", user)
}
\arguments{
\item{user}{}
}
\value{
db_conn
}
\description{
Connect to AML Project ORM Database
}

