% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_builders.r
\name{construct_rna_qc_dataframe}
\alias{construct_rna_qc_dataframe}
\title{Construct a data frame of RNA-SeQC results}
\usage{
construct_rna_qc_dataframe(exp.design)
}
\arguments{
\item{exp.design}{Experimental design data frame.}
}
\value{
df A data frame containing the RNA-SeQC results.
}
\description{
Given a data frame containing library names and RNA-SeQC paths, Construct a data frame of RNA-SeQC results
NOTE that this function makes strong assumptions about the shape of the incoming data frame.
}
