% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_builders.r
\name{construct_sailfish_gene_dataframe}
\alias{construct_sailfish_gene_dataframe}
\title{Construct a data frame of Sailfish gene-level results}
\usage{
construct_sailfish_gene_dataframe(exp.design)
}
\arguments{
\item{exp.design}{Experimental design data frame}
}
\value{
df A data frame containing the Sailfish results.
}
\description{
Given a data frame containing library names and Sailfish paths, Construct a data frame of Sailfish results
NOTE that this function makes strong assumptions about the shape of the incoming data frame,
and currently only fetches \emph{gene-level} results.
}
