% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_builders.r
\name{convert_sailfish_df_to_matrix}
\alias{convert_sailfish_df_to_matrix}
\title{From a data frame of Sailfish gene-level results, construct an expression matrix}
\usage{
convert_sailfish_df_to_matrix(sailfish.df, metric = c("NumReads", "TPM"))
}
\arguments{
\item{sailfish.df}{Data frame containing sailfish results}

\item{metric}{Expression metric to use (either 'NumReads' or 'TPM')}
}
\value{
df A data frame containing the Sailfish results
}
\description{
Given a data frame containing Sailfish gene-level results, from the construct_sailfish_gene_dataframe
function, convert that dataframe to a matrix suitable for downstream analysis
}

