% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{downregulated_genes}
\alias{downregulated_genes}
\title{Pull down-regulated genes}
\usage{
downregulated_genes(df, adjpval_threshold, log2fc_threshold)
}
\arguments{
\item{df}{Data frame containing DE results}

\item{adjpval_threshold}{Adjusted p-value threshold}

\item{log2fc_threshold}{Log2-fold-change threshold}
}
\value{

}
\description{
Pull down-regulated genes
}
