% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.r
\name{eng_log_breaks}
\alias{eng_log_breaks}
\title{Generate a set of breaks using the 'Engineer's log scale'}
\usage{
eng_log_breaks(x)
}
\arguments{
\item{x}{A vector of numeric data}
}
\value{
breaks A set of breaks suitable for ggplot breaks functions
}
\description{
Generate a set of breaks using the 'Engineer's log scale'
}
