% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{expression_summary}
\alias{expression_summary}
\title{Summarize distribution characteristics for a column of interest}
\usage{
expression_summary(df, x)
}
\arguments{
\item{df}{Data frame}

\item{x}{Bare name of column to summarize}
}
\value{

}
\description{
Summarize distribution characteristics for a column of interest
}
\examples{
df <- tibble(bar = rnorm(10000))
expression_summary(df, bar)
}
