% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.r
\name{fit_coxph}
\alias{fit_coxph}
\title{Fit a Cox proportional hazards model}
\usage{
fit_coxph(exp.design, label_col)
}
\arguments{
\item{exp.design}{Experimental design data frame}

\item{label_col}{Column in the design data frame used to separate samples}
}
\value{
cx The cox proportional hazards model
}
\description{
Fit a Cox proportional hazards model
}
