% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{ggsave_pdf_and_png}
\alias{ggsave_pdf_and_png}
\title{Save plot to PDF and PNG}
\usage{
ggsave_pdf_and_png(filename_root, plot, width, height, ...)
}
\arguments{
\item{filename_root}{Root filename (without extension)}

\item{plot}{ggplot object}

\item{width}{plot width in inches}

\item{height}{plot height in inches}

\item{...}{parameters passed to ggsave}

\item{height}{plot height in inches}
}
\description{
Save plot to PDF and PNG
}
