% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.r
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{Generate a grid of plots with a shared legend}
\usage{
grid_arrange_shared_legend(..., ncol = length(list(...)), nrow = 1,
  position = c("bottom", "right"))
}
\arguments{
\item{...}{A set of plots to arrange with a shared legend}
}
\value{
A ggplot plot
}
\description{
Adapted from:
  https://github.com/tidyverse/ggplot2/wiki/Share-a-legend-between-two-ggplot2-graphs
  http://rpubs.com/sjackman/grid_arrange_shared_legend
  http://stackoverflow.com/questions/13649473/add-a-common-legend-for-combined-ggplots
}
