% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{group_tally_table}
\alias{group_tally_table}
\title{Group / Tally / Table}
\usage{
group_tally_table(x, ...)
}
\arguments{
\item{x}{A data frame.}

\item{...}{Unquoted variables to group by.}
}
\value{
A knitr::kable().
}
\description{
Group / Tally / Table
}
