% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.r
\name{hit_plotter}
\alias{hit_plotter}
\title{Generate a dot-plot of expression of selected genes, facetted by selected feature}
\usage{
hit_plotter(expression_df, feature_label)
}
\arguments{
\item{expression_df}{Data frame containing expression values}

\item{feature_label}{Label in the data frame to facet by}
}
\value{
p A ggplot object containing the generated plot
}
\description{
Generate a dot-plot of expression of selected genes, facetted by selected feature
}

