% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_voom_for_feature.r
\name{limma_logFC_to_signed_foldchange}
\alias{limma_logFC_to_signed_foldchange}
\title{Convert limma's logFC value to signed fold-change}
\usage{
limma_logFC_to_signed_foldchange(logFC)
}
\arguments{
\item{logFC}{A logFC value reported by limma}
}
\value{
foldchange A signed fold-change value
}
\description{
Note that this needs to convert negative logFC numbers to absolute values to get correct fold-change estimates.
limma reports `logFC` as the 'estimate of the log2-fold-change'. For fold-changes in the \emph{negative} direction
we need to use the absolute value of the fold-change before applying the sign
}

