% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cna_workflow.r
\name{log2_plus1_transform}
\alias{log2_plus1_transform}
\title{Transform data by taking the log2(count+1)}
\usage{
log2_plus1_transform(x)
}
\arguments{
\item{x}{A matrix containing un-transformed expression (in some count format)}
}
\value{
A matrix with transformed values
}
\description{
Transform data by taking the log2(count+1)
}
\examples{
\dontrun{log2_plus1_transform(matrix(c(1, 2, 4, 8)))}
}
