% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_voom_for_feature.r
\name{munge_voom_to_ipa}
\alias{munge_voom_to_ipa}
\title{Munge differential expression hits in voom format to IPA-friendly format}
\usage{
munge_voom_to_ipa(voom_hits)
}
\arguments{
\item{voom_hits}{A data frame containing voom hits}
}
\value{
df A data frame containing the voom hits in IPA-friendly format
}
\description{
Munge differential expression hits in voom format to IPA-friendly format
}
