% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{num_to_text}
\alias{num_to_text}
\title{Convert a small number to a text representation}
\usage{
num_to_text(x, capitalize = FALSE)
}
\arguments{
\item{x}{An integer to convert to text}

\item{capitalize}{Boolean whether to capitalize text}
}
\description{
Convert a small number to a text representation
}
\examples{
num_to_text(5, capitalize = TRUE)
}
