% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifxparsers.r
\name{parse_pavfinder_0.3.0}
\alias{parse_pavfinder_0.3.0}
\title{Parse PAVfinder 0.3.0 output}
\usage{
parse_pavfinder_0.3.0(events_path)
}
\arguments{
\item{events_path}{Path to the PAVfinder events file}
}
\value{
A data frame with parsed PAVfincer results
}
\description{
Parse PAVfinder 0.3.0 output
}
