% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment_plot.r
\name{pathway_enrichment_plot}
\alias{pathway_enrichment_plot}
\title{Generate a plot of enriched pathways}
\usage{
pathway_enrichment_plot(gage_hits, plot_limit = 20)
}
\arguments{
\item{gage_hits}{A data frame produced by the gage function}

\item{plot_limit}{The number of top pathways to plot}
}
\value{
p A ggplot object containing the generated plot
}
\description{
Generate a plot of enriched pathways
}

