% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_helpers.r
\name{plot_rf_predictions}
\alias{plot_rf_predictions}
\title{Plot predictions from a randomForest model}
\usage{
plot_rf_predictions(rf_predictions)
}
\arguments{
\item{rf_predictions}{A vector of predictions from a randomForest model}
}
\value{
p A ggplot plot
}
\description{
Plot predictions from a randomForest model
}
