% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.r
\name{plot_survival}
\alias{plot_survival}
\title{Plot survival curves}
\usage{
plot_survival(exp.design, label_col)
}
\arguments{
\item{exp.design}{Experimental design data frame}

\item{label_col}{Column in the design data frame used to separate samples}
}
\value{
p A ggplot object containing the generated plot
}
\description{
Plot survival curves
}

