% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{pretty_num}
\alias{pretty_num}
\title{Set a function for pretty-printing numbers in the text}
\usage{
pretty_num(x, ...)
}
\arguments{
\item{x}{Number to print}

\item{...}{Additional options to be passed to prettyNum}
}
\value{
A formatted number
}
\description{
Set a function for pretty-printing numbers in the text
}
\examples{
pretty_num(5.55654)
}
