% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_builders.r
\name{pull_enrichr_enriched_terms}
\alias{pull_enrichr_enriched_terms}
\title{Pull EnrichR enriched terms for a single pathway source}
\usage{
pull_enrichr_enriched_terms(
  enrichr_lst,
  pathway_source,
  cohort_name,
  padj_thresh
)
}
\arguments{
\item{enrichr_lst}{Result list returned from EnrichR}

\item{pathway_source}{Name of pathway source to query}

\item{cohort_name}{Cohort name to apply}

\item{padj_thresh}{Adjusted p-value threshold}
}
\value{
Data frame of enriched pathways.
}
\description{
Pull EnrichR enriched terms for a single pathway source
}
