% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_builders.r
\name{pull_enrichr_terms_to_plot}
\alias{pull_enrichr_terms_to_plot}
\title{Identify enriched terms across cohorts}
\usage{
pull_enrichr_terms_to_plot(long_df)
}
\arguments{
\item{long_df}{Long dataframe containing enrichment results across cohorts}
}
\value{

}
\description{
Identify enriched terms across cohorts
}
