% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_helpers.r
\name{random_forest_summary}
\alias{random_forest_summary}
\title{From a random forest model generated by caret, generate summary plots}
\usage{
random_forest_summary(caret.rf, yvals)
}
\arguments{
\item{caret.rf}{A randomForest model generated by caret}

\item{yvals}{Feature in the experimental design data frame to use as data labels}
}
\value{
caret.forest.importance A data frame containing variable importance for the random forest model
}
\description{
From a random forest model generated by caret, generate summary plots
}
