% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv_indel_helpers.r
\name{read_filled_variant_tsv}
\alias{read_filled_variant_tsv}
\title{This function reads in the 'long' TSV variant sheets generated by the makefile}
\usage{
read_filled_variant_tsv(tsv_file)
}
\arguments{
\item{tsv_file}{A TSV-formatted file containing SNV data.}
}
\value{
filled.df A data frame containing filled SNV data
}
\description{
This function reads in the 'long' TSV variant sheets generated by the makefile
}
