% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.r
\name{retrieve_db_comments}
\alias{retrieve_db_comments}
\title{Retrieve DB Comments}
\usage{
retrieve_db_comments(db_conn)
}
\arguments{
\item{db_conn}{Database connection object}
}
\value{
db_comments.df A data-frame containing curated result comments from the database
}
\description{
Retrieve DB Comments
}

