% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.r
\name{retrieve_db_curated_results}
\alias{retrieve_db_curated_results}
\title{Retrieve DB Curated Results
\lifecycle{deprecated}}
\usage{
retrieve_db_curated_results(db_conn)
}
\arguments{
\item{db_conn}{Database connection object}
}
\value{
db_curated_results.df A data-frame containing curated result information from the database
}
\description{
Retrieve DB Curated Results
\lifecycle{deprecated}
}
