% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.r
\name{retrieve_db_libraries}
\alias{retrieve_db_libraries}
\title{Retrieve DB Libraries}
\usage{
retrieve_db_libraries(db_conn)
}
\arguments{
\item{db_conn}{Database connection object}
}
\value{
db_libraries.df A data-frame containing library information from the database
}
\description{
Retrieve DB Libraries
}
