% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.r
\name{retrieve_db_paths}
\alias{retrieve_db_paths}
\title{Retrieve DB Paths}
\usage{
retrieve_db_paths(db_conn)
}
\arguments{
\item{db_conn}{Database connection object}
}
\value{
db_paths.df A data-frame containing path information from the database
}
\description{
Retrieve DB Paths
}

