% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.r
\name{retrieve_db_specimens}
\alias{retrieve_db_specimens}
\title{Retrieve DB Specimens
\lifecycle{deprecated}}
\usage{
retrieve_db_specimens(db_conn)
}
\arguments{
\item{db_conn}{Database connection object}
}
\value{
db_specimens.df A data-frame containing specimen information from the database
}
\description{
Retrieve DB Specimens
\lifecycle{deprecated}
}
