% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_predictions.r
\name{rf_predictions}
\alias{rf_predictions}
\title{Extract predictions from a random forest model}
\usage{
rf_predictions(caret.rf, yvals)
}
\arguments{
\item{caret.rf}{A randomForest model generated by caret}

\item{yvals}{Feature in the experimental design data frame to use as data labels}
}
\value{
rf.predict A vector of predictions from the randomForest model
}
\description{
Extract predictions from a random forest model
}

