% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifxparsers.r
\name{rna_seqc_parser}
\alias{rna_seqc_parser}
\title{Parse output from RNA-SeQC v1.1.8}
\usage{
rna_seqc_parser(rna_seqc_path)
}
\arguments{
\item{rna_seqc_path}{file}
}
\value{
df
}
\description{
Parse output from RNA-SeQC v1.1.8
}
\examples{
\dontrun{qc.dat <- rna_seqc_parser(rna_seqc_path)}
}
