% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cna_workflow.r
\name{run_cna_workflow}
\alias{run_cna_workflow}
\title{Run the CNA workflow}
\usage{
run_cna_workflow(expr.mat, genes.bed, window_size = NA,
  mean_tpm_threshold = 0, cv_threshold = 0, reference_labels,
  scale_samples_by_zero = TRUE)
}
\arguments{
\item{scale_samples_by_zero}{}
}
\description{
Run the CNA workflow
}
