% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gage_for_feature.r
\name{run_gage_for_feature}
\alias{run_gage_for_feature}
\title{Run GAGE gene-set enrichment analysis for a given feature}
\usage{
run_gage_for_feature(voom.hits, gsets, same.dir = TRUE)
}
\arguments{
\item{voom.hits}{A data frame from the voom analysis, _unthresholded_}

\item{gsets}{A list of gene sets, constructed by the geneIds function above}

\item{same.dir}{same.dir argument for the `gage` function}
}
\value{
pathways.merged A data frame containing up- and down-regulated pathways
}
\description{
Run GAGE gene-set enrichment analysis for a given feature
}

