% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_voom_for_feature.r
\name{run_voom_for_feature}
\alias{run_voom_for_feature}
\title{Run voom for a given feature comparison}
\usage{
run_voom_for_feature(yvals, exp.design, counts.mat)
}
\arguments{
\item{yvals}{Feature in the experimental design data frame to use as data labels}

\item{exp.design}{Experimental design data frame}

\item{counts.mat}{Matrix containing count data}
}
\value{
df A data frame containing the differential expression scores for all genes
}
\description{
Run voom for a given feature comparison
}

