% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifxparsers.r
\name{sailfish_gene_parser_post_0.7.0}
\alias{sailfish_gene_parser_post_0.7.0}
\title{Parse gene-level output from Sailfish >=0.7.0}
\usage{
sailfish_gene_parser_post_0.7.0(sailfish_path)
}
\arguments{
\item{sailfish_path}{file}
}
\value{
df
}
\description{
Parse gene-level output from Sailfish >=0.7.0
}
\examples{
sailfish.genes.df <- sailfish_gene_parser_post_0.7.0(sailfish_path)
}
