% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{select_by_position}
\alias{select_by_position}
\title{Quick function for mapping to select out list items}
\usage{
select_by_position(x, position)
}
\arguments{
\item{x}{A vector}

\item{position}{The position to select out}
}
\description{
Quick function for mapping to select out list items
}
\examples{
select_by_position(lst, 1)
}
