% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv_indel_helpers.r
\name{split_gatk_format_vals}
\alias{split_gatk_format_vals}
\title{Split a FORMAT string from GATK-haplotype to get simpler statistics}
\usage{
split_gatk_format_vals(df)
}
\arguments{
\item{df}{}
}
\value{
updated_df
}
\description{
Split a FORMAT string from GATK-haplotype to get simpler statistics
}
\examples{
\dontrun{split_gatk_format_vals(df)}
}
