% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv_indel_helpers.r
\name{spread_filled_snv_df_to_wide}
\alias{spread_filled_snv_df_to_wide}
\title{This function takes the 'long, filled' data frames with variant observations, and converts them to 'wide' format}
\usage{
spread_filled_snv_df_to_wide(df)
}
\arguments{
\item{df}{A data frame containing SNV data in the 'long, filled' format}
}
\value{
df A wide-format data frame containing simplified SNV data
}
\description{
This function takes the 'long, filled' data frames with variant observations, and converts them to 'wide' format
}
