% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cna_workflow.r
\name{subset_genes_by_thresholds}
\alias{subset_genes_by_thresholds}
\title{Subset expression matrix by thresholds}
\usage{
subset_genes_by_thresholds(
  expr.mat,
  mean_tpm_threshold,
  cv_threshold,
  reference_means,
  reference_cvs
)
}
\arguments{
\item{expr.mat}{A matrix of expression values}

\item{mean_tpm_threshold}{The mean TPM threshold. Genes with values above this threshold will be retained}

\item{cv_threshold}{The coefficient of variation threshold. Genes with values below this threshold will be retained}

\item{reference_means}{The reference means, generated by the calculate_reference_set_mean_expression function}

\item{reference_cvs}{The reference CVs, generated by the calculate_reference_set_coefficient_of_variation function}
}
\value{
expr.mat
}
\description{
Subset expression matrix by thresholds
}
