% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_genes_for_plots.r
\name{subset_genes_for_plots}
\alias{subset_genes_for_plots}
\title{Subset the TPM matrix to the genes of interest}
\usage{
subset_genes_for_plots(tpm.mat, exp.design, gene_list)
}
\arguments{
\item{tpm.mat}{Matrix containing expression values}

\item{exp.design}{Experimental design data frame}

\item{gene_list}{Vector of gene names to extract}
}
\value{
hits_to_plot_long A data frame containing hits for plotting
}
\description{
Subset the TPM matrix to the genes of interest
}
