% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_heatmap.r
\name{subset_heatmap}
\alias{subset_heatmap}
\title{Generate a heatmap for a subset of genes}
\usage{
subset_heatmap(gene_set, annotation_col, counts.mat)
}
\arguments{
\item{gene_set}{A vector containing gene names to plot}

\item{annotation_col}{Data frame containing annotation columns for pheatmap}

\item{counts.mat}{A matrix of gene count data}
}
\value{
p A ggplot plot
}
\description{
Generate a heatmap for a subset of genes
}

