% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.r
\name{subset_pc_plot}
\alias{subset_pc_plot}
\title{Generate a prinicple component plot for a subset of genes, coloured by a given feature}
\usage{
subset_pc_plot(gene_set, exp.design, feature_label, counts.mat)
}
\arguments{
\item{gene_set}{A vector containing gene names to subset}

\item{exp.design}{Experimental design data frame}

\item{feature_label}{Label in the data frame to facet by}

\item{counts.mat}{Matrix containing expression count data}
}
\value{
p A ggplot object containing the generated plot
}
\description{
Generate a prinicple component plot for a subset of genes, coloured by a given feature
}
