% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.r
\name{subtype_to_disease}
\alias{subtype_to_disease}
\title{From a subtype indication, return a simplified disease type}
\usage{
subtype_to_disease(subtype)
}
\arguments{
\item{subtype}{Curated disease subtype}
}
\value{
disease A simplified disease name
}
\description{
From a subtype indication, return a simplified disease type
}
