% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.r
\name{surv_panel_plot}
\alias{surv_panel_plot}
\title{Modified survival plot}
\usage{
surv_panel_plot(surv.fit, data, palette, title, pval.coord = c(600, 0.9),
  xlim = c(0, 800), ...)
}
\arguments{
\item{surv.fit}{A survival object from survival::Surv}

\item{data}{The data frame used to generate the survival object}

\item{palette}{A named list containing plotting colours}

\item{title}{Plot title}

\item{...}{Additional parameters to pass to ggsurvplot}
}
\description{
Modified survival plot
}
