% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{tfl_split_sort}
\alias{tfl_split_sort}
\title{A function to sort TFL IDs into ascending order}
\usage{
tfl_split_sort(df)
}
\arguments{
\item{df}{}
}
\description{
A function to sort TFL IDs into ascending order
}
