% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_helpers.r
\name{train_caret_rf}
\alias{train_caret_rf}
\title{Train a random forest model with caret}
\usage{
train_caret_rf(gene_set, yvals, exp.design, counts.mat)
}
\arguments{
\item{gene_set}{A vector containing gene names to subset}

\item{yvals}{Feature in the experimental design data frame to use as data labels}

\item{exp.design}{Experimental design data frame}

\item{counts.mat}{A matrix of gene count data}
}
\value{
caret.rf A an object containing a caret Randomforest
}
\description{
Train a random forest model with caret
}

