% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plotter.r
\name{volcano_plotter}
\alias{volcano_plotter}
\title{Generate a volcano plot of expression hits}
\usage{
volcano_plotter(voom_hits, label_threshold = 30)
}
\arguments{
\item{voom_hits}{Data frame containing voom differential expression results}

\item{label_threshold}{Threshold for -log10(PVal) to add a label}
}
\value{
p A ggplot object containing the generated plot
}
\description{
Generate a volcano plot of expression hits
}

