% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.r
\name{connect_to_aml_project_orm}
\alias{connect_to_aml_project_orm}
\title{Connect to AML Project ORM Database}
\usage{
connect_to_aml_project_orm(
  dbname = "aml_project_db",
  host = "karsanlab-db01",
  user
)
}
\arguments{
\item{dbname}{Database name}

\item{host}{Database host}

\item{user}{Database username}
}
\value{
db_conn A database connection object
}
\description{
Connect to AML Project ORM Database
}
