% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{downregulated_genes}
\alias{downregulated_genes}
\title{downregulated_genes}
\usage{
downregulated_genes(df, adjpval_threshold, log2fc_threshold)
}
\arguments{
\item{log2fc_threshold}{}
}
\value{

}
\description{
downregulated_genes
}
