% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cna_workflow.r
\name{log2_plus1_transform}
\alias{log2_plus1_transform}
\title{Transform data by taking the log2(TPM+1)}
\usage{
log2_plus1_transform(tpm.mat)
}
\arguments{
\item{expr.mat}{A matrix containing untransformed expression (in transcripts per million)}
}
\value{
tpm.log2.mat A matrix with transformed values
}
\description{
Transform data by taking the log2(TPM+1)
}
\examples{
tpm.log2.mat <- log2_plus1_transform(tpm.mat)
}
