% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cna_workflow.r
\name{long_df_to_ranked_matrix}
\alias{long_df_to_ranked_matrix}
\title{Transform long dataframe to sorted matrix}
\usage{
long_df_to_ranked_matrix(long.df)
}
\arguments{
\item{long.df}{}
}
\value{
ret.mat
}
\description{
Transform long dataframe to sorted matrix
}
