% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{matrix_glimpse}
\alias{matrix_glimpse}
\title{# Report basic information about a matrix}
\usage{
matrix_glimpse(mat)
}
\arguments{
\item{mat}{A matrix}
}
\description{
# Report basic information about a matrix
}
\examples{
matrix_glimpse(matrix(runif(1000), nrow = 200, ncol = 500))
}
