% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifxparsers.r
\name{parse_pavfinder_0.3.0}
\alias{parse_pavfinder_0.3.0}
\title{Parse pavfinder 0.2.0 output}
\usage{
parse_pavfinder_0.3.0(events_path)
}
\arguments{
\item{events_path}{Path to the PAVfinder events file}
}
\value{

}
\description{
Parse pavfinder 0.2.0 output
}
