% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv_indel_helpers.r
\name{plot_vars_by_coverage}
\alias{plot_vars_by_coverage}
\title{This function takes the long and wide data frames, and plots all variants by covearge, colouring them by concordance status}
\usage{
plot_vars_by_coverage(wide.df, long.df, sample_name)
}
\arguments{
\item{sample_name}{}
}
\description{
This function takes the long and wide data frames, and plots all variants by covearge, colouring them by concordance status
}

