% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv_indel_helpers.r
\name{plot_vars_by_vaf}
\alias{plot_vars_by_vaf}
\title{Join to the plotting frame and re-plot for VAF}
\usage{
plot_vars_by_vaf(wide.df, long.df, sample_name)
}
\arguments{
\item{wide.df}{A wide dataframe containing variant observations}

\item{long.df}{A long dataframe containing variant observations}

\item{sample_name}{Sample name}
}
\value{
A plot
}
\description{
Join to the plotting frame and re-plot for VAF
}
