% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gage_for_feature.r
\name{run_gage_for_feature}
\alias{run_gage_for_feature}
\title{Plot predictions from a randomForest model}
\usage{
run_gage_for_feature(voom.hits, gsets, adj_p_threshold = 0.05,
  same.dir = TRUE)
}
\arguments{
\item{voom.hits}{A data frame from the voom analysis, _unthresholded_}

\item{adj_p_threshold}{Threshold cutoff to use for voom adjusted p-value}

\item{same.dir}{same.dir argument for the `gage` function}

\item{gene_list}{A list of gene sets, constructed by the geneIds function above}
}
\value{
pathways.merged A data frame containing up- and down-regulated pathways
}
\description{
Plot predictions from a randomForest model
}

