% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cna_workflow.r
\name{scale_by_rolling_mean}
\alias{scale_by_rolling_mean}
\title{Scale expression estimates by rolling mean}
\usage{
scale_by_rolling_mean(expr.mat, genes.bed, window_size)
}
\arguments{
\item{expr.mat}{a matrix of expression values}

\item{genes.bed}{A bed-like file containing gene coordinates}

\item{window_size}{Window size to use for rolling mean}
}
\value{
A transformed matrix
}
\description{
Scale expression estimates by rolling mean
}
