% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv_indel_helpers.r
\name{spread_filled_snv_df_to_wide}
\alias{spread_filled_snv_df_to_wide}
\title{This function takes the 'long, filled' data frames with variant observations, and converts them to 'wide' format}
\usage{
spread_filled_snv_df_to_wide(df)
}
\arguments{
\item{df}{}
}
\value{
df
}
\description{
This function takes the 'long, filled' data frames with variant observations, and converts them to 'wide' format
}
