% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{upregulated_genes}
\alias{upregulated_genes}
\title{upregulated_genes}
\usage{
upregulated_genes(df, adjpval_threshold, log2fc_threshold)
}
\arguments{
\item{log2fc_threshold}{}
}
\description{
upregulated_genes
}
